Куча всего

+++ app/Http/Controllers/API/ChatController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API;
 use App\Enums\Role;
 use App\Http\Controllers\Controller;
 use App\Http\Library\ApiHelpers;
+use App\Http\Requests\StoreChatMemberRequest;
 use App\Http\Requests\StoreChatRequest;
 use App\Http\Requests\StoreMessageRequest;
 use App\Http\Requests\UpdateChatRequest;
@@ -29,15 +30,19 @@ class ChatController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request, $userId) : JsonResponse
+    public function index(Request $request) : JsonResponse
     {
-        if (empty($userId) && Role::isAdminFromValue($request->user()->role)) {
+        $this->authorize('viewAllUserChat', Chat::class);
+
+        $user = $request->user();
+
+        if (Role::isAdmin($user)) {
             $chats = Chat::all();
         }
-        else{
-            $chats = Chat::query()->whereExists(function ($query) use ($userId) {
-                $query->select(DB::raw(1))->from('chat_members')->where('user_id', $userId);
-            });
+        else {
+            $chats = Chat::query()->whereExists(function ($query) use ($user) {
+                $query->select(DB::raw(1))->from('chat_members')->where('user_id', $user->id);
+            })->get();
         }
 
         return $this->onSuccess(new ChatCollection($chats), 'Chats retrieved successfully.');
@@ -48,6 +53,8 @@ class ChatController extends Controller
      */
     public function store(StoreChatRequest $request) : JsonResponse
     {
+        $this->authorize('create', Chat::class);
+
         $chat = Chat::create($request->validated());
         ChatMember::create(['chat_id' => $chat->id, 'user_id' => $request->user()->id]);
         return $this->onSuccess(new ChatResource($chat), 'Chat created successfully.', 201);
@@ -56,17 +63,22 @@ class ChatController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Request $request, Chat $chat) : JsonResponse
+    public function show(Chat $chat) : JsonResponse
     {
-        if (ChatMember::where('chat_id', $chat->id)
+        $this->authorize('view', $chat);
+
+        return $this->onSuccess(new ChatResource($chat), 'Chat retrieved successfully.');
+        /*
+        if ($chat->chatMembers()
                 ->where('user_id', $request->user()->id)
                 ->exists()
-            || Role::isAdminFromValue($request->user()->role))
+            || Role::isAdmin($request->user()))
         {
             return $this->onSuccess(new ChatResource($chat), 'Chat retrieved successfully.');
         }
 
         return $this->onError(401, 'You are not allowed to see this chat.');
+        */
     }
 
     /**
@@ -74,6 +86,7 @@ class ChatController extends Controller
      */
     public function update(UpdateChatRequest $request, Chat $chat) : JsonResponse
     {
+        $this->authorize('update', $chat);
         $chat->update($request->validated());
         return $this->onSuccess(new ChatResource($chat), 'Chat updated successfully.');
     }
@@ -83,6 +96,7 @@ class ChatController extends Controller
      */
     public function destroy(Chat $chat) : JsonResponse
     {
+        $this->authorize('delete', $chat);
         $chat->delete();
         return $this->onSuccess(null, 'Chat deleted successfully.');
     }
@@ -90,6 +104,7 @@ class ChatController extends Controller
 
     public function getMessages(Request $request, Chat $chat) : JsonResponse
     {
+        //TODO: Подумать, в какую политику запихнуть это
         if (ChatMember::where('chat_id', $chat->id)
             ->where('user_id', $request->user()->id)
             ->exists()) {
@@ -100,19 +115,27 @@ class ChatController extends Controller
 
         return $this->onError(401, 'You do not have permission to view this chat.');
     }
-    public function sendMessage(StoreMessageRequest $request) : JsonResponse
+    public function sendMessage(StoreMessageRequest $request, Chat $chat) : JsonResponse
     {
+        $this->authorize('create', [Message::class, $chat]);
         $data = $request->validated();
-        $message = Chat::find($data['chat_id'])->messages()->create($data);
+        $data['sender_id'] = $request->user()->id;
+        $message = $chat->messages()->create($data);
         return $this->onSuccess(new MessageResource($message), 'Message sent successfully.', 201);
     }
     public function updateMessage(UpdateMessageRequest $request, Message $message) : JsonResponse
     {
+        $this->authorize('update', Message::class);
         $message->update($request->validated());
         return $this->onSuccess(new MessageResource($message), 'Message updated successfully.');
     }
-    public function deleteMessage(Request $request, Message $message) : JsonResponse
+    public function deleteMessage(Message $message) : JsonResponse
     {
+        $this->authorize('delete', $message);
+        $message->delete();
+        return $this->onSuccess(null, 'Message deleted successfully.');
+
+        /*
         if ($request->user()->id === $message->sender_id)
         //|| Role::isAdmin($request->user()->role))
         {
@@ -121,48 +144,30 @@ class ChatController extends Controller
         }
 
         return $this->onError(401, 'You do not have permission to delete this chat.');
+        */
     }
 
     public function getMembers(Request $request, Chat $chat) : JsonResponse
     {
-        if (ChatMember::where('chat_id', $chat->id)
-            ->where('user_id', $request->user()->id)
-            ->exists()
-        || Role::isAdminFromValue($request->user()->role)) {
-            return $this->onSuccess(new UserCollection($chat->chatMembers()->get()));
-        }
-
-        return $this->onError(401, 'You do not have permission to view this chat.');
+        $this->authorize('view', [ChatMember::class, $chat]);
+        return $this->onSuccess(new UserCollection($chat->chatMembers()->get()));
     }
-    public function addMember(Request $request, Chat $chat, User $user) : JsonResponse
+    public function addMember(StoreChatMemberRequest $request, Chat $chat) : JsonResponse
     {
-        //TODO: сделать отдельный StoreMemberRequest
-
-        if (ChatMember::where('chat_id', $chat->id)
-            ->where('user_id', $request->user()->id)
-            ->exists()) {
-
-            //TODO: Посмотреть про guard аттрибуты
-            ChatMember::firstOrCreate(['user_id' => $user->id, 'chat_id' => $chat->id]);
-            return $this->onSuccess(null, 'Chat member added successfully.', 201);
-        }
-        return $this->onError(401, 'You are not allowed to use this action.');
+        $this->authorize('create', [ChatMember::class, $request->user(), $chat]);
+        $user_id = $request->validated()['user_id'];
+        ChatMember::firstOrCreate(['user_id' => $user_id, 'chat_id' => $chat->id]);
+        return $this->onSuccess(null, 'Chat member added successfully.', 201);
     }
     public function removeMember(Request $request, Chat $chat, User $user) : JsonResponse
     {
-        if (Role::isAdminFromValue($request->user()->role) ||
-            ($request->user()->id === $user->id && ChatMember::firstWhere([
-                    ['user_id' => $user->id],
-                    ['chat_id' => $chat->id]
-                ])->exists()))
-        {
-            ChatMember::firstWhere([
-                ['user_id' => $user->id],
-                ['chat_id' => $chat->id]
-            ])->delete();
-            return $this->onSuccess(null, 'Chat member removed successfully.');
-        }
+        $chatMember = ChatMember::firstWhere([
+            ['user_id' => $user->id],
+            ['chat_id' => $chat->id]
+        ]);
+        $this->authorize('delete', [ChatMember::class, $request->user(), $chatMember]);
 
-        return $this->onError(401, 'You do not have permission to view this chat.');
+        $chatMember->delete();
+        return $this->onSuccess(null, 'Chat member removed successfully.');
     }
 }
